fix: fix cluster resource pool selection [DET-7517 DET-7567 DET-7517] (#4391)
- deprecate the old cluster page and associated components.
- remove the logic around persisting user RP selection.
- fix an issue where the jobs displayed in the job table would be
for a previously selected resource pool.
