+++ docs/release-notes/4391-cluster-overview.txt
@@ -1,0 +1,11 @@
+:orphan:
+
+**Breaking Changes**
+
+-  WebUI: Remove previously unlisted cluster page. This page has been replaced by a new version
+      available through the navigation bar.
+
+**Bug Fixes**
+
+-  WebUI: Fix a bug where a previous resource pool selection would not update when a new
+      resource pool is selected for viewing associated jobs.
