--- webui/react/src/components/ResourcePoolCard.module.scss
@@ -1,120 +1,0 @@
-$padding: 12px;
-$icon-side: 40px;
-$line-height: 22px;
-
-.rp-type-logo {
-  width: $icon-side;
-}
-.base {
-  background-color: var(--theme-stage);
-  border-radius: var(--theme-border-radius);
-  box-shadow: var(--theme-elevation);
-  color: var(--theme-stage-on);
-  display: flex;
-  flex-direction: column;
-  font-size: 14px;
-  line-height: $line-height;
-  overflow: hidden;
-  text-overflow: ellipsis;
-
-  .body {
-    display: flex;
-    flex-direction: column;
-    flex-grow: 1;
-  }
-  .header {
-    background: var(--theme-stage);
-    color: white;
-    display: flex;
-    flex-direction: row;
-    min-height: 80px;
-    padding: $padding;
-
-    .icon {
-      align-items: center;
-      background-color: var(--theme-stage-strong);
-      border: var(--theme-stroke-width) solid var(--theme-stage-border);
-      border-radius: 50%;
-      display: flex;
-      flex-shrink: 0;
-      height: $icon-side;
-      justify-content: center;
-      margin: auto $padding auto 0;
-      width: $icon-side;
-    }
-    .info {
-      margin: auto 0;
-
-      .name {
-        margin-bottom: 5px;
-      }
-    }
-    .tags *[class*='Badge'] {
-      margin-bottom: 5px;
-      margin-right: 10px;
-    }
-    .tags *[class*='Badge']:last-child {
-      margin-right: 0;
-    }
-  }
-  .agentsStatus {
-    background: var(--theme-status-inactive);
-    color: white;
-    display: flex;
-    font-size: 12px;
-    height: 24px;
-    text-align: center;
-    vertical-align: middle;
-
-    & > * {
-      margin: auto;
-    }
-  }
-  .agentsStatus.active {
-    background-color: var(--theme-status-active);
-  }
-  .empty {
-    color: var(--theme-stage-on-weak);
-  }
-  .description {
-    align-content: center;
-    display: flex;
-    flex-direction: column;
-    height: $line-height * 4;
-    padding: 0;
-    position: relative;
-
-    p {
-      margin: auto 0;
-      // https://css-tricks.com/line-clampin/
-      overflow: auto;
-      padding: $padding;
-    }
-  }
-  hr {
-    background-color: var(--theme-stage-border);
-    border: none;
-    height: var(--theme-stroke-width);
-    margin: 0;
-  }
-  section {
-    padding: $padding;
-  }
-  section.details {
-    display: flex;
-    flex-direction: column;
-    flex-grow: 1;
-
-    ul {
-      flex-grow: 1;
-    }
-  }
-  section > *:not(:first-child) {
-    padding-top: $padding;
-  }
-}
-.spaceBetweenHorizontal {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-}
