--- webui/react/src/pages/Cluster/ClusterOverview.module.scss
@@ -1,16 +1,0 @@
-.base {
-  .descriptionColumn {
-    max-width: calc(max(50vw, 300px));
-    word-break: break-word;
-    word-wrap: break-word;
-  }
-  .chartColumn {
-    min-width: 150px;
-  }
-  :global(label.ant-radio-button-wrapper) {
-    font-size: inherit;
-    height: inherit;
-    line-height: inherit;
-    padding: 4px;
-  }
-}
