+++ webui/react/src/pages/Cluster/utils.test.ts
@@ -1,6 +1,6 @@
 import { ResourcePool } from 'types';
 
-import { maxPoolSlotCapacity } from './ClusterOverview';
+import { maxPoolSlotCapacity } from '../Clusters/ClustersOverview';
 
 const pools: Record<string, Partial<ResourcePool>> = {
   dynamic: { maxAgents: 3, slotsAvailable: 0, slotsPerAgent: 2 },
